requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
click==7.0                # via pip-tools
django-csp==3.5
django-vanilla-views==1.0.6
django-webpack-loader==0.6.0
django==2.2
pip-tools==3.6.1
pytz==2019.1              # via django
six==1.12.0               # via pip-tools
sqlparse==0.3.0           # via django
